requirements.txt
# ── LLM & Intelligence ──────────────────────────────────────────────────────────
google-genai                       # Gemini (latest stable SDK — Replaces legacy google-generativeai)
Pillow>=10.0.0                     # Image handling for Gemini Vision

# ── API & Orchestration ────────────────────────────────────────────────────────
fastapi[standard]                  # FastAPI + Uvicorn + Multipart
pydantic                           # Data validation

# ── Databases & Persistence ────────────────────────────────────────────────────
psycopg2-binary>=2.9.9             # Postgres driver
neo4j>=6.1.0                       # Neo4j driver (Stable — Replaces deprecated neo4j-driver)
sqlalchemy>=2.0.0                  # SQL engine
pandas>=2.1.0                      # Data structures

# ── Universal Ingestion ────────────────────────────────────────────────────────
lxml>=6.0.2                        # High-performance XML/HTML (required for exact line hinting)
openpyxl>=3.1.0                    # .xlsx support
xlrd>=2.0.1                        # .xls support (legacy)
pdfplumber>=0.11.0                 # PDF text extraction
pdf2image>=1.17.0                  # PDF → image (multimodal fallback)

# ── Connectors ─────────────────────────────────────────────────────────────────
requests>=2.31.0                   # REST calls
simple-salesforce>=1.12.5          # Salesforce API

# ── Testing & Validation ───────────────────────────────────────────────────────
pytest==9.0.2                      # High-performance testing framework
httpx==0.28.1                       # Async HTTP client for integration tests
pytest-cov==7.0.0                  # Test coverage enforcement
pytest-mock==3.15.1                # Atomic mocking for AI/DB substrates
pytest-asyncio==1.3.0               # Async test support
